+++ app/src/main/java/myapplication/android/pixelpal/data/source/platform/PlatformRemoteSource.kt
@@ -5,7 +5,7 @@ import myapplication.android.pixelpal.data.models.platforms.PlatformsList
 
 interface PlatformRemoteSource {
 
-    suspend fun getPlatformDetails(id: Int): PlatformDetails
+    suspend fun getPlatformDetails(id: Long): PlatformDetails
 
     suspend fun getPlatforms(page: Int): PlatformsList
